+++ src/components/Home/Skills/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { FC, useRef, useState, lazy, Suspense } from 'react';
+import React, { FC,  useState, lazy, Suspense } from 'react';
 
 import BgParticles from './BgParticles';
 import SkillGrouped from './SkillGrouped';
@@ -10,16 +10,13 @@ import { useLocale } from 'next-intl';
 import { useTranslations } from 'next-intl';
 import { AnimatePresence } from 'framer-motion';
 
-import { HiOutlineUserGroup } from 'react-icons/hi';
-import { HiOutlineDesktopComputer, HiOutlineDatabase } from 'react-icons/hi';
-
-import Tab, { TabItem } from '@/components/UI/Tab';
 import HeaderArticle from '@/components/Common/HeaderArticle';
 
 import { trpcReact as trpc } from '@/utils/trpc';
 import { LIMIT_PER_PAGE_BIG } from '@/utils/constants';
 
 import type { SkillType } from '@/utils/interfaces/types';
+import FilterType from './FilterType';
 
 const SkillModal = lazy(() => import('./SkillModal'));
 
@@ -58,55 +55,35 @@ const Skills: FC<ISkillsProps> = ({}) => {
   return (
     <>
       <div className="relative bg-background-50">
-        <BgParticles />
-        <HeaderArticle showIcon title={t('title')} subtitle={t('subtitle')} description={t('description')} />
-        <article id="skills" className="z-10 mx-auto overflow-hidden px-4 lg:container lg:px-20 pb-4 lg:pb-20">
-          <div className="py-10 relative z-10 lg:grid lg:grid-cols-12 lg:gap-16 lg:items-center">
-            <div className="lg:col-span-5">
-              <div className="relative">
-                <Tab
-                  variant="secondary"
-                  currentTab={value}
-                  setCurrentTab={setValue}
-                  className="flex flex-col gap-4 z-20">
-                  <TabItem
-                    icon={HiOutlineDesktopComputer}
-                    title={t('tabs.frontend.title')}
-                    description={t('tabs.frontend.description')}
-                  />
-                  <TabItem 
-                    icon={HiOutlineDatabase} 
-                    title={t('tabs.backend.title')} 
-                    description={t('tabs.backend.description')}
-                  />
-                  <TabItem
-                    icon={HiOutlineUserGroup}
-                    title={t('tabs.tools.title')}
-                    description={t('tabs.tools.description')}
-                  />
-                </Tab>
+        <article className="mx-auto px-4 lg:container lg:px-20 pb-4 lg:pb-20">
+          <BgParticles />
+          <HeaderArticle showIcon title={t('title')} subtitle={t('subtitle')} description={t('description')} />
+          <div id="skills" className="z-10 overflow-hidden">
+            <div className="mb-10 lg:mb-0 relative z-10 lg:grid lg:grid-cols-12 lg:gap-16 lg:items-center">
+              <div className="lg:col-span-5">
+                <FilterType value={value} onChange={setValue} />
+              </div>
+              <div className="lg:col-span-7">
+                <SkillGrouped
+                  onClick={handleOpenModal}
+                  type={value === 0 ? 'frontend' : value === 1 ? 'backend' : 'devops' }
+                  loading={value === 0 ? frontend?.isLoading : value === 1 ? backend?.isLoading : tools?.isLoading}
+                  skills={value === 0 ? frontend?.data?.data || [] : value === 1 ? backend?.data?.data || [] : tools?.data?.data || []}
+                />
               </div>
             </div>
-            <div className="lg:col-span-7">
-              <SkillGrouped
-                onClick={handleOpenModal}
-                type={value === 0 ? 'frontend' : value === 1 ? 'backend' : 'devops' }
-                loading={value === 0 ? frontend?.isLoading : value === 1 ? backend?.isLoading : tools?.isLoading}
-                skills={value === 0 ? frontend?.data?.data || [] : value === 1 ? backend?.data?.data || [] : tools?.data?.data || []}
-              />
-            </div>
-          </div>
 
-          <div className="container mx-auto flex w-full justify-center py-5">
-            <Link
-              scroll={true}
-              href="/certificates"
-              className="group z-30 mt-4 inline-flex items-center gap-1 text-sm font-medium text-primary-600">
-              {t('modal.seeCertificates')}
-              <span aria-hidden="true" className="block transition group-hover:translate-x-0.5">
-                →
-              </span>
-            </Link>
+            <div className="container mx-auto flex w-full justify-center py-5">
+              <Link
+                scroll={true}
+                href="/certificates"
+                className="group z-30 mt-4 inline-flex items-center gap-1 text-sm font-medium text-primary-600">
+                {t('modal.seeCertificates')}
+                <span aria-hidden="true" className="block transition group-hover:translate-x-0.5">
+                  →
+                </span>
+              </Link>
+            </div>
           </div>
         </article>
       </div>