+++ CHANGELOG.md
@@ -9,6 +9,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - Improve i18n configuration
 - Update Particles.js to latest version
 - Fix theme preview in the theme selector component and add Christmas theme
+- Improve skills section with new design and new data structure
 
 ## 2.5.2 (2023-08-14)